+++ src/Commands/Remote/SSHBaseCommand.php
@@ -3,12 +3,14 @@
 namespace Pantheon\Terminus\Commands\Remote;
 
 use Pantheon\Terminus\Commands\TerminusCommand;
+use Pantheon\Terminus\Exceptions\TerminusProcessException;
 use Pantheon\Terminus\Helpers\LocalMachineHelper;
 use Pantheon\Terminus\Models\Environment;
 use Pantheon\Terminus\Models\Site;
 use Pantheon\Terminus\Site\SiteAwareInterface;
 use Pantheon\Terminus\Site\SiteAwareTrait;
-use Pantheon\Terminus\Exceptions\TerminusProcessException;
+use Symfony\Component\Console\Output\ConsoleOutputInterface;
+use Symfony\Component\Process\Process;
 use Symfony\Component\Process\ProcessUtils;
 
 /**
@@ -32,6 +34,10 @@ abstract class SSHBaseCommand extends TerminusCommand implements SiteAwareInterf
      * @var Site
      */
     private $site;
+    /**
+     * @var bool
+     */
+    protected $progressAllowed;
 
     /**
      * Define the environment and site properties
@@ -41,6 +47,20 @@ abstract class SSHBaseCommand extends TerminusCommand implements SiteAwareInterf
     protected function prepareEnvironment($site_env_id)
     {
         list($this->site, $this->environment) = $this->getSiteEnv($site_env_id);
+    }
+
+    /**
+     * progressAllowed sets the field that controls whether a progress bar
+     * may be displayed when a program is executed. If allowed, a progress
+     * bar will be used in tty mode.
+     *
+     * @param type|bool $allowed
+     * @return $this
+     */
+    protected function setProgressAllowed($allowed = true)
+    {
+        $this->progressAllowed = $allowed;
+        return $this;
     }
 
     /**
@@ -84,7 +104,8 @@ abstract class SSHBaseCommand extends TerminusCommand implements SiteAwareInterf
         }
         return $this->getContainer()->get(LocalMachineHelper::class)->execute(
             $ssh_command,
-            $this->getOutputCallback()
+            $this->getOutputCallback(),
+            $this->progressAllowed
         );
     }
 
@@ -196,8 +217,14 @@ abstract class SSHBaseCommand extends TerminusCommand implements SiteAwareInterf
     {
         if ($this->getContainer()->get(LocalMachineHelper::class)->useTty() === false) {
             $output = $this->output();
-            return function ($type, $buffer) use ($output) {
-                $output->write($buffer);
+            $stderr = $this->stderr();
+
+            return function ($type, $buffer) use ($output, $stderr) {
+                if (Process::ERR === $type) {
+                    $stderr->write($buffer);
+                } else {
+                    $output->write($buffer);
+                }
             };
         }
         return function ($type, $buffer) {