+++ src/Commands/Remote/WPCommand.php
@@ -24,13 +24,17 @@ class WPCommand extends SSHBaseCommand
      *
      * @param string $site_env Site & environment in the format `site-name.env`
      * @param array $wp_command WP-CLI command
+     * @param array $options Commandline options
+     * @option progress Allow progress bar to be used (tty mode only)
      * @return string Command output
      *
      * @usage <site>.<env> -- <command> Runs the WP-CLI command <command> remotely on <site>'s <env> environment.
+     * @usage <site>.<env> --progress -- <command> Runs a WP-CLI command with a progress bar
      */
-    public function wpCommand($site_env, array $wp_command)
+    public function wpCommand($site_env, array $wp_command, array $options = ['progress' => false])
     {
         $this->prepareEnvironment($site_env);
+        $this->setProgressAllowed($options['progress']);
         return $this->executeCommand($wp_command);
     }
 }