+++ tests/unit_tests/Helpers/LocalMachineHelperTest.php
@@ -70,51 +70,11 @@ class LocalMachineHelperTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests the LocalMachineHelper::execute($command, $callback) function when interactive
+     * Tests the LocalMachineHelper::execute($command, $callback, $progress) function
      */
-    public function testExecuteWhenInteractive()
+    public function testExecute()
     {
-        $command = 'ls';
-        $callback = function () {
-            return null;
-        };
-
-        $this->input->method('isInteractive')->willReturn(true);
-        $this->config->expects($this->once())
-            ->method('get')
-            ->with('timeout')
-            ->willReturn(55);
-
-        $output = $this->local_machine->execute($command, $callback);
-        $this->assertEquals(0, $output['exit_code']);
-    }
-
-    /**
-     * Tests the LocalMachineHelper::execute($command, $callback) function when not interactive
-     */
-    public function testExecuteWhenNotInteractive()
-    {
-        $command = 'ls';
-        $callback = function () {
-            return null;
-        };
-
-        $this->input->method('isInteractive')->willReturn(false);
-        $this->config->expects($this->once())
-            ->method('get')
-            ->with('timeout')
-            ->willReturn(55);
-
-        $output = $this->local_machine->execute($command, $callback);
-        $this->assertEquals(0, $output['exit_code']);
-    }
-
-    /**
-     * Tests the LocalMachineHelper::execInteractive($command, $callback) function
-     */
-    public function testExecInteractive()
-    {
-        $out = $this->local_machine->execInteractive('ls');
+        $out = $this->local_machine->execute('ls');
         $this->assertEquals(0, $out['exit_code']);
     }